\name{DataScienceInterviewPractice-package}
\alias{DataScienceInterviewPractice-package}
\alias{DataScienceInterviewPractice}
\docType{package}
\title{
\packageTitle{DataScienceInterviewPractice}
}
\description{
\packageDescription{DataScienceInterviewPractice}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DataScienceInterviewPractice}
\packageIndices{DataScienceInterviewPractice}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{DataScienceInterviewPractice}

Maintainer: \packageMaintainer{DataScienceInterviewPractice}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
